fixed orm

+++ src/main/java/br/com/gama/wipro/repositories/CurrentAccountRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.gama.wipro.entities.CurrentAccount;
 
-public interface CurrentAccountRepository extends JpaRepository<CurrentAccount, Long>{
+public interface CurrentAccountRepository extends JpaRepository<CurrentAccount, Integer>{
 
 }
